Backend/JobScrape/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
lxml
httpx
pydantic
python-dotenv
flask
flask_cors
# For secure & dynamic scraping
playwright

# (Optional) If you want browser control via Selenium instead of Playwright
# selenium
# webdriver-manager

# For async features (if needed in scraper)
anyio

# Enable CORS (cross-origin requests)
fastapi[all]
psycopg2-binary

